% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesdept.R
\name{buttonkrmesdeptodsmatadataUI_right}
\alias{buttonkrmesdeptodsmatadataUI_right}
\title{预览区域}
\usage{
buttonkrmesdeptodsmatadataUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
