% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesroute.R
\name{buttonkrmesrouteodsmatadataUI_left}
\alias{buttonkrmesrouteodsmatadataUI_left}
\title{操作区域}
\usage{
buttonkrmesrouteodsmatadataUI_left()
}
\value{
返回值
}
\description{
操作区域
}
\examples{
buttonUI_left()
}
